absolute_cinema/requirements.txt
# FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Modelos de dados
pydantic==2.5.0

# YouTube API
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Análise de sentimentos
textblob==0.17.1

# Tradução
deep-translator==1.11.4

# Variáveis de ambiente
python-dotenv==1.0.0

# CORS
python-multipart==0.0.6

# Utilitários
requests==2.31.0
